# docs/api-spec/invoice-billing-service-openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: billing@example.com
    name: Finance Systems Team
  description: |
    Microservice for generating invoices, recording payments, and chasing overdue balances.
    Supports draft and issued invoices, credit notes, and tax breakdowns.
  title: Invoice Billing Service
  version: 1.0.0
servers:
- description: Production
  url: https://api.example.com/v1
- description: Local development
  url: http://localhost:8080/v1
...
paths:
  /invoices:
    get:
      description: "Returns a paginated list of invoices, optionally filtered by status\
        \ or customer. Supports standard pagination parameters (page, size)."
      operationId: listInvoices
      parameters:
      - description: Filter by the billed customer
        explode: true
        in: query
        name: customerId
        required: false
        schema:
          type: string
        style: form
      - description: Filter by invoice status
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - DRAFT
          - ISSUED
          - PAID
          - OVERDUE
          - VOID
          type: string
        style: form
      - description: Page number (starts at 1)
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Number of items per page
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
          description: Successful response
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request - the server could not understand the payload
            or parameters.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error - something went wrong on the server
            side.
      summary: List all invoices
      tags:
      - Invoices
      x-accepts: application/json
      x-tags:
      - tag: Invoices
    post:
      description: Creates a draft invoice from a set of line items. The `invoiceNumber`
        is generated by the server and is unique.
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
          description: Invoice created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request - the server could not understand the payload
            or parameters.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: An invoice already exists for this billing period
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error - something went wrong on the server
            side.
      summary: Create a new invoice
      tags:
      - Invoices
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Invoices
  /invoices/{invoiceId}:
    delete:
      description: Voids the invoice so it no longer counts towards receivables. Paid
        invoices cannot be voided.
      operationId: voidInvoice
      parameters:
      - description: Unique invoice identifier (UUID)
        explode: false
        in: path
        name: invoiceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Invoice voided successfully (no content returned)
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The requested resource was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error - something went wrong on the server
            side.
      summary: Void an invoice
      tags:
      - Invoices
      x-accepts: application/json
      x-tags:
      - tag: Invoices
    get:
      description: Retrieves the full details of a single invoice including line items
        and tax breakdown.
      operationId: getInvoiceById
      parameters:
      - description: Unique invoice identifier (UUID)
        explode: false
        in: path
        name: invoiceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
          description: Invoice found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invoice not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error - something went wrong on the server
            side.
      summary: Get an invoice by ID
      tags:
      - Invoices
      x-accepts: application/json
      x-tags:
      - tag: Invoices
    put:
      description: Replaces the editable fields of a draft invoice. Issued invoices
        can only have their due date changed.
      operationId: updateInvoice
      parameters:
      - description: Unique invoice identifier (UUID)
        explode: false
        in: path
        name: invoiceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
          description: Invoice updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request - the server could not understand the payload
            or parameters.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The requested resource was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error - something went wrong on the server
            side.
      summary: Update an existing invoice
      tags:
      - Invoices
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Invoices
components:
...
  schemas:
    Invoice:
      example:
        dueDate: 2026-08-31T00:00:00.000+00:00
        lineItems:
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        totalAmount: 1416.0
        createdAt: 2026-08-01T09:00:00Z
        amountPaid: 0
        subtotal: 1200.0
        invoiceNumber: INV-2026-00317
        customerId: cust-88213
        currency: USD
        id: i1n2v3c4
        issuedAt: 2026-08-01T00:00:00.000+00:00
        taxAmount: 216.0
        status: ISSUED
      properties:
        id:
          description: Unique invoice ID (UUID)
          example: i1n2v3c4
          type: string
        invoiceNumber:
          description: Sequential invoice number (unique)
          example: INV-2026-00317
          type: string
        customerId:
          example: cust-88213
          type: string
        status:
          enum:
          - DRAFT
          - ISSUED
          - PAID
          - OVERDUE
          - VOID
          example: ISSUED
          type: string
        currency:
          example: USD
          type: string
        subtotal:
          example: 1200.0
          format: double
          type: number
        taxAmount:
          example: 216.0
          format: double
          type: number
        totalAmount:
          example: 1416.0
          format: double
          type: number
        amountPaid:
          example: 0
          format: double
          type: number
        lineItems:
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
          type: array
        issuedAt:
          example: 2026-08-01
          format: date
          type: string
        dueDate:
          example: 2026-08-31
          format: date
          type: string
        createdAt:
          example: 2026-08-01T09:00:00Z
          format: date-time
          type: string
      required:
      - customerId
      - id
      - invoiceNumber
      - status
      - totalAmount
      type: object
    InvoiceLineItem:
      example:
        unitPrice: 100.0
        quantity: 12
        taxRatePercent: 18.0
        description: Platform subscription - August 2026
      properties:
        description:
          example: Platform subscription - August 2026
          type: string
        quantity:
          example: 12
          minimum: 1
          type: integer
        unitPrice:
          example: 100.0
          format: double
          type: number
        taxRatePercent:
          example: 18.0
          format: double
          type: number
      required:
      - description
      - quantity
      - unitPrice
      type: object
    CreateInvoiceRequest:
      example:
        lineItems:
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        purchaseOrderRef: PO-4471
        dueDate: 2026-08-31T00:00:00.000+00:00
        customerId: cust-88213
        currency: USD
      properties:
        customerId:
          example: cust-88213
          type: string
        lineItems:
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
          minItems: 1
          type: array
        currency:
          default: USD
          example: USD
          type: string
        dueDate:
          example: 2026-08-31
          format: date
          type: string
        purchaseOrderRef:
          example: PO-4471
          type: string
      required:
      - customerId
      - dueDate
      - lineItems
      type: object
    UpdateInvoiceRequest:
      example:
        lineItems:
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        - unitPrice: 100.0
          quantity: 12
          taxRatePercent: 18.0
          description: Platform subscription - August 2026
        purchaseOrderRef: PO-4488
        dueDate: 2026-09-15T00:00:00.000+00:00
      properties:
        dueDate:
          example: 2026-09-15
          format: date
          type: string
        purchaseOrderRef:
          example: PO-4488
          type: string
        lineItems:
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
          type: array
      type: object
    InvoiceListResponse:
      example:
        size: 20
        totalPages: 16
        page: 1
        content:
        - dueDate: 2026-08-31T00:00:00.000+00:00
          lineItems:
          - unitPrice: 100.0
            quantity: 12
            taxRatePercent: 18.0
            description: Platform subscription - August 2026
          - unitPrice: 100.0
            quantity: 12
            taxRatePercent: 18.0
            description: Platform subscription - August 2026
          totalAmount: 1416.0
          createdAt: 2026-08-01T09:00:00Z
          amountPaid: 0
          subtotal: 1200.0
          invoiceNumber: INV-2026-00317
          customerId: cust-88213
          currency: USD
          id: i1n2v3c4
          issuedAt: 2026-08-01T00:00:00.000+00:00
          taxAmount: 216.0
          status: ISSUED
        - dueDate: 2026-08-31T00:00:00.000+00:00
          lineItems:
          - unitPrice: 100.0
            quantity: 12
            taxRatePercent: 18.0
            description: Platform subscription - August 2026
          - unitPrice: 100.0
            quantity: 12
            taxRatePercent: 18.0
            description: Platform subscription - August 2026
          totalAmount: 1416.0
          createdAt: 2026-08-01T09:00:00Z
          amountPaid: 0
          subtotal: 1200.0
          invoiceNumber: INV-2026-00317
          customerId: cust-88213
          currency: USD
          id: i1n2v3c4
          issuedAt: 2026-08-01T00:00:00.000+00:00
          taxAmount: 216.0
          status: ISSUED
        totalElements: 317
      properties:
        page:
          example: 1
          type: integer
        size:
          example: 20
          type: integer
        totalElements:
          example: 317
          type: integer
        totalPages:
          example: 16
          type: integer
        content:
          items:
            $ref: "#/components/schemas/Invoice"
          type: array
      type: object
    ErrorResponse:
      properties:
        status:
          example: 400
          type: integer
        errorCode:
          example: INVALID_DUE_DATE
          type: string
        message:
          example: "Validation failed: dueDate must be in the future"
          type: string
        timestamp:
          example: 2026-08-12T10:00:00Z
          format: date-time
          type: string
        path:
          example: /v1/invoices
          type: string
      required:
      - message
      - status
      type: object